Data Store/requirements.txt
APScheduler==3.6.3
apscheduler-bundle==1.1.2
connexion==2.5.1
DBUtils==1.3
Flask==1.1.1
Flask-Cors==3.0.8
mysql==0.0.2
mysql-connector-python==8.0.19
mysqlclient==1.4.4
pykafka==2.8.0
PyMySQL==0.9.3
PyYAML==5.1
request==2019.3.22
requests==2.20.1
SQLAlchemy==1.3.8
swagger-ui-bundle==0.0.6